requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dateutil>=2.8.0

